Move items-related stuffs to item mod from visitor mod
Move `rewrite_extern_crate`, is_mod_decl`, `is_use_item` and `is_extern_crate`.
